Add login form and template page
The new form is defined in the forms file, and it inherits from
FlaskForm in the flask_wtf package. This external packege makes it easy
to create new forms by only definig the type of input that the form will
handle.

A new login view function has been added to the routes file. This view
shows a blank login form when the user first requests this page,
validates the data submitted and redirects to the index page if
everything is correct.

The login template renders a simple form based on the class variables
defined in LoginForm. If the form has any invalid fields when submitted,
it shows all errors.