Add DB support and User model
The new packages added to the project (SQLAlquemy and flask-migrate),
will allow us to create tables in our database based in the schemas
defined in the models file. The flask-migrate packege will be important
when we need to update or revert the database schema without affecting
the data already present.

The new User model adds a table to the database that will represent each
user on our site.