Add login/logout functionality
The app uses the logic defined in the flask-login package to allow users
to log in and out of the web site. This library also takes care of
checking if the user is already logged in, and of keeping some pages
private until the user authenticates.

The index view is now private--only authenticated users can see it. On
the other hand, the login view now truly verifies that the credentials
of the user exist and are correct before logging them in. A new view to
log out the user has been added, and uses the default functionality
provided by flask-login.

In the models file, the User model implements the required methods for
flask-login by means of the UserMixin provided to the class definition.
Also, the new function load_user is necessary for flask-login to know
which user, if any, is currently logged in.