Add register page
The new register template renders the registration form and also shows
to the user any validation errors, so they can correct them.

A link to the registration form has been added at the end of the login
form, so user that don't have an account can register.

The new register view takes care of actually creating the new user,
sending any errors to the template and redirecting the user to the login
form once the user has registered successfully.