Add profile page for users
Authenticated users can now see their profile page or the profile page
of other users. This page will show all the posts of the user, but
currently it only shows some mock posts.

The user view function makes sure that the user for the requested profile
page exists, otherwise it shows a 404 error.

There is a new link in the base template that redirects to the profile
page of the current user.