Add functionality to edit profile information
A new form has been added that will allow the user to change their
username and/or about me information. This form is displayed with
help of the edit_profile template, which also shows any validation
errors.

The new view function in the routes file checks that the information
submitted on the form is valid. If the form is valid, it stores the new
data in the user table, otherwise it renders any validation errors.

The before_request function has been moved to the start of the file.

The profile page now displays the link to edit the profile information,
but only if the user is the owner of the account.