Add handle for 404 and 500 HTTP errors
The errors file handles what will happen when a 404 or 500 error occur
in the application, and render the appropriate template to the user.

The templates for the 404 and 500 errors are very similar. Both notify
the user of the error and have a link for the user to click to go back
to their home page.