Fix duplicate username bug
The form to edit the user profile now correctly checks that the new
username being requested doesn't exist before attempting to update
the User record.

This is not a perfect solution, because it may not work when two or more
processes are accessing the database at the same time, causing a race
condition, but it's good enough for now.