Add form to create posts
A new form has been added to the index route to allow users to create
new posts. The form includes a text area for the post content and a
submit button. Upon submission, the post is saved to the database and
the user is redirected back to the index page where the new post is
displayed.

The index page now displays posts from all users the current user is
following instead of mock data.