Add pagination of posts in user profile
The posts in the user profile page are now paginated to improve
the user experience when viewing a large number of posts. This change
includes updates to the routes and templates to handle pagination logic.

The total number of posts that can be displayed on each page has been
set to 20.