Update styling in form helpers
The logic used to add icons to form fields has been updated be simpler
to understand and maintain. The new logic uses a dictionary to map
field types to their corresponding icons, making it easier to add or
change icons in the future.

The logic to create the class of the form field has been simplified
as well. Instead of using multiple if-else statements, a single line
of code now constructs the class string based on the field's properties.
The form macro can also add autofocus to the first field of the form if
specified.

The styling of error messages has been improved to make them more
noticeable. When a field has errors, the error messages are displayed
in a red text below the field, and the field itself is highlighted
with a red border.

Password fields now have a toggle button to show or hide the password.