Added a better UI for profile editing
The form to edit the profile has been redesigned to improve user
experience, using the simple_form template to keep consistency across
the application.

The fields of the EditProfileForm have new names to be more descriptive.

A minor change was made to the routes to redirect users to their profile
page after editing it.