Remove minified JS files no longer needed and update Dockerfile
The minified versions of the JavaScript files have been removed as they
are no longer necessary. All JavaScript files are minified using terser
during the build process of the docker image.

The Dockerfile has been updated to ensure that the preline.js file is
copied to a permanent location before the node_modules directory is
deleted, ensuring that the application can still access this file after
the build process.

The route to serve the preline.js file has been updated so that it
serves the file from its new location based on the configuration
variable PRELINE_JS_DIR.