Add viewing for user messages
There is a new route to view user messages, a link in the navigation bar
to access it, and a new template to display the messages. The existing
post display template has been modified to handle both posts and
messages.

New translations have been added for the new text elements.

I also added type hints to some routes that didn't have them yet.